The official text format isn't for inline assembly.
Current compilers aren't planning to use the official text format for
inline assembly and .S files. And, the debugging use case is already
mentioned in the previous paragraph, and we can mention profiling there
too. So this paragraph just needs to mention that the part about
browsers not accepting the text format in regular content.