Move push/pop into a seperate document for now.
It's still worth pursuing, but it's unfamiliar and somewhat complex.
Taking it out for now will allow us to focus the discussion on the
overall approach of this text format, which is the more important
discussion right now. We can revisit push/pop in the future if this
overall approach works.